Add ProjetilSapo
botei projetil no sapo ( é configuravel a vel do atk, o intervalo do dano e o dano)

+++ SwampDefense/Assets/Scripts/Enemys/EnemyLife.cs
@@ -2,9 +2,9 @@ using UnityEngine;
 
 public class MoscaVida : MonoBehaviour
 {
-    public int vida = 100;
+    public float vida = 100f;
 
-    public void ReceberDano(int dano)
+    public void ReceberDano(float dano)
     {
         vida -= dano;
